Intent is unclear
```sh
if ! "${SCRIPT_PATH}" ; then
```
This way to execute a script and to test its execution is a bit unusual (at least to me). At first, it was unclear to me whether this line tests the variable (set or not), the file (exists or not) or the execution (succeeds or not). To execute the script and print a message if execution fails, I suggest to use a `command || echo "fail"`.

Passing the script to a bash subshell saves a `chmod +x`. The subshell should inherit the export values, but I didn't tested it.